requirements.txt
attrs==24.2.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
distlib==0.3.9
et_xmlfile==2.0.0
filelock==3.16.1
h11==0.14.0
identify==2.6.3
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.1.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
pycodestyle==2.12.1
pyflakes==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.28.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
